# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/awaitpayment": {
            "get": {
                "description": "Hiển thị phiên đấu giá chưa thanh toán của người dùng. Cần phải đăng nhập để sử dụng. Trả về JSON form",
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Hiển thị phiên đấu giá chưa thanh toán",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/history/bid": {
            "get": {
                "description": "Hiển thị những hoạt động đấu giá của user trên các session.",
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Hiển thị lịch sử đấu giá của user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BidHistory"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/history/sell": {
            "get": {
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Hiển thị lịch sử bán hàng của user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SellHistory"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/lock/:sessionid": {
            "put": {
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Khóa phiên đấu giá (API nội bộ, không public cho user)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session ID",
                        "name": "SessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/session": {
            "get": {
                "description": "Tìm phiên đấu giá bằng tên mặt hàng và/hoặc categories. Tên mặt hàng không cần viết đầy đủ. Trả về tất cả nếu để trống. Trả về JSON form",
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Tìm phiên đấu giá bằng tên mặt hàng và/hoặc categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the item (or part of it)",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Item Categories by number",
                        "name": "categories",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SessionSearch"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Thay đổi thông tin phiên đấu giá.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Information to be provided",
                        "name": "NewSessionInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateSession"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Tạo phiên đấu giá mới",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Information to be provided",
                        "name": "NewSessionInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NewSession"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/session/:sessionid": {
            "get": {
                "description": "Tìm phiên đáu giá bằng id dưới dạng path, trả về JSON form",
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Tìm phiên đấu giá bằng id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SessionSearch"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa toàn bộ phiên đấu giá và các dữ liệu liên quan như lịch sử đấu giá, thông tin mặt hàng của phiên đấu giá.",
                "tags": [
                    "BidSession Controller"
                ],
                "summary": "Xóa phiên đấu giá (Administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "session id to be deleted",
                        "name": "sessionid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.BidHistory": {
            "type": "object",
            "properties": {
                "bidamount": {
                    "type": "integer"
                },
                "biddate": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "itemid": {
                    "type": "integer"
                },
                "itemname": {
                    "type": "string"
                },
                "sessionenddate": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                },
                "sessionstartdate": {
                    "type": "string"
                }
            }
        },
        "model.BidSessionLog": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createAt": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "model.SellHistory": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "itemid": {
                    "type": "integer"
                },
                "itemname": {
                    "type": "string"
                },
                "sessionenddate": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                },
                "sessionstartdate": {
                    "type": "string"
                }
            }
        },
        "model.SessionSearch": {
            "type": "object",
            "required": [
                "categoriesid",
                "currentbid",
                "itemcondition",
                "itemdescription",
                "itemname",
                "minimumincreasebid"
            ],
            "properties": {
                "biddingLog": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BidSessionLog"
                    }
                },
                "categoriesid": {
                    "type": "integer"
                },
                "categoriesname": {
                    "type": "string"
                },
                "currentbid": {
                    "type": "integer"
                },
                "enddate": {
                    "type": "string"
                },
                "imagelink": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "itemcondition": {
                    "type": "string"
                },
                "itemdescription": {
                    "type": "string"
                },
                "itemid": {
                    "type": "integer"
                },
                "itemname": {
                    "type": "string"
                },
                "minimumincreasebid": {
                    "type": "integer"
                },
                "sellerid": {
                    "type": "string"
                },
                "sellername": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                },
                "sessionstatus": {
                    "type": "string"
                },
                "startdate": {
                    "type": "string"
                },
                "viewcount": {
                    "type": "integer"
                }
            }
        },
        "model.UpdateSession": {
            "type": "object",
            "required": [
                "sessionid"
            ],
            "properties": {
                "itemcondition": {
                    "type": "string"
                },
                "itemdescription": {
                    "type": "string"
                },
                "itemname": {
                    "type": "string"
                },
                "sessionid": {
                    "type": "integer"
                }
            }
        },
...
    }
}